refactor: global response 부분 위치 변경
