+++ backend/src/main/java/consolelog/auth/exception/LoginFailedException.java
@@ -1,6 +1,6 @@
 package consolelog.auth.exception;
 
-import consolelog.global.error.ErrorCode;
+import consolelog.global.response.ErrorCode;
 import consolelog.global.exception.BusinessException;
 
 public class LoginFailedException extends BusinessException {
