+++ backend/src/main/java/consolelog/comment/exception/CommentNotFoundException.java
@@ -2,7 +2,7 @@ package consolelog.comment.exception;
 
 
 import consolelog.global.exception.BusinessException;
-import consolelog.global.error.ErrorCode;
+import consolelog.global.response.ErrorCode;
 
 public class CommentNotFoundException extends BusinessException {
