+++ backend/src/main/java/consolelog/comment/exception/ReplyDepthException.java
@@ -1,6 +1,6 @@
 package consolelog.comment.exception;
 
-import consolelog.global.error.ErrorCode;
+import consolelog.global.response.ErrorCode;
 import consolelog.global.exception.BusinessException;
 
 public class ReplyDepthException extends BusinessException {
