+++ backend/src/main/java/consolelog/global/response/ResultCode.java
@@ -1,4 +1,4 @@
-package consolelog.global.result;
+package consolelog.global.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
