+++ backend/src/main/java/consolelog/global/support/token/TokenNotFoundException.java
@@ -1,7 +1,7 @@
 package consolelog.global.support.token;
 
 import consolelog.global.exception.UnauthorizedException;
-import consolelog.global.error.ErrorCode;
+import consolelog.global.response.ErrorCode;
 
 public class TokenNotFoundException extends UnauthorizedException {
