+++ backend/src/main/java/consolelog/image/exception/S3UploadException.java
@@ -1,7 +1,7 @@
 package consolelog.image.exception;
 
 import consolelog.global.exception.BusinessException;
-import consolelog.global.error.ErrorCode;
+import consolelog.global.response.ErrorCode;
 
 public class S3UploadException extends BusinessException {
     public S3UploadException() {
