+++ backend/src/main/java/consolelog/like/controller/LikeController.java
@@ -1,7 +1,7 @@
 package consolelog.like.controller;
 
 import consolelog.auth.dto.AuthInfo;
-import consolelog.global.result.ResultResponse;
+import consolelog.global.response.ResultResponse;
 import consolelog.global.support.token.Login;
 import consolelog.like.dto.LikeFlipResponse;
 import consolelog.like.service.LikeService;
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import static consolelog.global.result.ResultCode.LIKE_CREATED_SUCCESS;
+import static consolelog.global.response.ResultCode.LIKE_CREATED_SUCCESS;
 
 
 @Tag(name = "Like", description = "Like API Document")
