+++ backend/src/main/java/consolelog/like/exception/CommentLikeNotFoundException.java
@@ -2,7 +2,7 @@ package consolelog.like.exception;
 
 
 import consolelog.global.exception.BusinessException;
-import consolelog.global.error.ErrorCode;
+import consolelog.global.response.ErrorCode;
 
 public class CommentLikeNotFoundException extends BusinessException {
