+++ backend/src/main/java/consolelog/member/exception/InvalidPasswordFormatException.java
@@ -1,6 +1,6 @@
 package consolelog.member.exception;
 
-import consolelog.global.error.ErrorCode;
+import consolelog.global.response.ErrorCode;
 import consolelog.global.exception.BusinessException;
 
 public class InvalidPasswordFormatException extends BusinessException {
