+++ backend/src/main/java/consolelog/member/exception/MemberNotFoundException.java
@@ -1,7 +1,7 @@
 package consolelog.member.exception;
 
 import consolelog.global.exception.BusinessException;
-import consolelog.global.error.ErrorCode;
+import consolelog.global.response.ErrorCode;
 
 public class MemberNotFoundException extends BusinessException {
