+++ backend/src/main/java/consolelog/member/exception/PasswordConfirmationException.java
@@ -1,6 +1,6 @@
 package consolelog.member.exception;
 
-import consolelog.global.error.ErrorCode;
+import consolelog.global.response.ErrorCode;
 import consolelog.global.exception.BusinessException;
 
 public class PasswordConfirmationException extends BusinessException {
