+++ backend/src/main/java/consolelog/post/exception/InvalidTitleException.java
@@ -1,6 +1,6 @@
 package consolelog.post.exception;
 
-import consolelog.global.error.ErrorCode;
+import consolelog.global.response.ErrorCode;
 import consolelog.global.exception.BusinessException;
 
 public class InvalidTitleException extends BusinessException {
