+++ backend/src/main/java/consolelog/post/exception/PostNotFoundException.java
@@ -1,7 +1,7 @@
 package consolelog.post.exception;
 
 import consolelog.global.exception.BusinessException;
-import consolelog.global.error.ErrorCode;
+import consolelog.global.response.ErrorCode;
 
 public class PostNotFoundException extends BusinessException {
