+++ backend/src/test/java/consolelog/auth/controller/AuthControllerTest.java
@@ -15,7 +15,7 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
 
-import static consolelog.global.result.ResultCode.*;
+import static consolelog.global.response.ResultCode.*;
 import static consolelog.util.fixture.AuthFixture.*;
 import static org.mockito.BDDMockito.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
